Crate objc_exception

Rust interface for Objective-C’s @throw and @try/@catch statements.

Enums

Exception

An opaque type representing any Objective-C object thrown as an exception.

Functions

throw⚠

Throws an Objective-C exception. The argument must be a pointer to an Objective-C object.

try⚠

Tries to execute the given closure and catches an Objective-C exception if one is thrown.